requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configobj==5.0.9
cryptography==44.0.2
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
idna==3.10
lxml==5.3.1
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pathlib==1.0.1
pdfplumber==0.11.5
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sentence-transformers==3.4.1
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
